fixes #3312 Host Deletes on backend but with errors in log
Former-commit-id: 5c54226a815dc0d766bf7c5ffa8b5d64445c709a

+++ src/com/dotmarketing/menubuilders/RefreshMenus.java
@@ -266,20 +266,22 @@ public class RefreshMenus {
 					Logger.error(RefreshMenus.class, e.getMessage(), e);
 					throw new DotRuntimeException(e.getMessage(), e);
 				} 
-				if(host == null){
-					Logger.error(RefreshMenus.class, "Folder id :" + auxFolder.getInode() + " has no host");
-					return;
-				}
 				java.io.File[] files = directory.listFiles();
-				for (int i = 0; i < files.length; i++) 
-				{
-					File file = files[i];
-					String fileName = file.getName();				
-					if(fileName.startsWith(host.getIdentifier()))
-					{
-						file.delete();
+				if(files.length > 0){
+					if(host == null){
+						Logger.error(RefreshMenus.class, "Folder id :" + auxFolder.getInode() + " has no host");
+						return;
 					}				
-				}				
+					for (int i = 0; i < files.length; i++) 
+					{
+						File file = files[i];
+						String fileName = file.getName();				
+						if(fileName.startsWith(host.getIdentifier()))
+						{
+							file.delete();
+						}				
+					}
+				}
 			}			
 		}
 		//http://jira.dotmarketing.net/browse/DOTCMS-1873
